+++ src/frontend/apps/impress/src/features/docs/doc-export/blocks-mapping/index.ts
@@ -6,6 +6,8 @@ export * from './headingPDF';
 export * from './imageDocx';
 export * from './imagePDF';
 export * from './paragraphPDF';
+export * from './pdfDocx';
+export * from './pdfPdf';
 export * from './quoteDocx';
 export * from './quotePDF';
 export * from './tablePDF';